hana/requirements.txt
colorama==0.4.4
cycler==0.11.0
Deprecated==1.2.13
fonttools==4.30.0
hana-ml==2.11.22020900
hdbcli==2.11.20
htmlmin==0.1.12
kiwisolver==1.3.2
loguru==0.6.0
matplotlib==3.5.1
numpy==1.21.5
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
pydot==1.4.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
schedule==1.1.0
Shapely==1.8.1.post1
six==1.16.0
tqdm==4.63.0
win32-setctime==1.1.0
wrapt==1.14.0
